Extracted serde code for Rect into separate file
